+++ CHANGES.rst
@@ -1,3 +1,27 @@
+0.16.0 (2020-12-18)
+-------------------
+New Features
+^^^^^^^^^^^^
+
+- Added an option to `to_fits_sip()` to be able to specify the reference
+  point (``crpix``) of the FITS WCS. [#337]
+
+- Added support for providing custom range of degrees in ``to_fits_sip``. [#339]
+
+Bug Fixes
+^^^^^^^^^
+
+- ``bounding_box`` now works with tuple of ``Quantities``. [#331]
+
+- Fix a formula for estimating ``crpix`` in ``to_fits_sip()`` so that ``crpix``
+  is near the center of the bounding box. [#337]
+
+- Allow sub-pixel sampling of the WCS model when computing SIP approximation in
+  ``to_fits_sip()``. [#338]
+
+- Fixed a bug in ``to_fits_sip`` due to which ``inv_degree`` was ignored. [#339]
+
+
 0.15.0 (2020-11-13)
 -------------------
 New Features